moved cfg_format into two files
